// movie-store-web-app/src/app/api/api-reference.ts
export class Customer implements ICustomer {
    id?: string;
    name?: string;
    email?: string;
    status?: Status;
    role?: Role;
    statusExpirationDate?: Date | undefined;
    moneySpent?: number;
    purchasedMovies?: PurchasedMovie[];

    constructor(data?: ICustomer) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.id = _data["id"];
            this.name = _data["name"];
            this.email = _data["email"];
            this.status = _data["status"];
            this.role = _data["role"];
            this.statusExpirationDate = _data["statusExpirationDate"] ? new Date(_data["statusExpirationDate"].toString()) : <any>undefined;
            this.moneySpent = _data["moneySpent"];
            if (Array.isArray(_data["purchasedMovies"])) {
                this.purchasedMovies = [] as any;
                for (let item of _data["purchasedMovies"])
                    this.purchasedMovies!.push(PurchasedMovie.fromJS(item));
            }
        }
    }

    static fromJS(data: any): Customer {
        data = typeof data === 'object' ? data : {};
        let result = new Customer();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["id"] = this.id;
        data["name"] = this.name;
        data["email"] = this.email;
        data["status"] = this.status;
        data["role"] = this.role;
        data["statusExpirationDate"] = this.statusExpirationDate ? this.statusExpirationDate.toISOString() : <any>undefined;
        data["moneySpent"] = this.moneySpent;
        if (Array.isArray(this.purchasedMovies)) {
            data["purchasedMovies"] = [];
            for (let item of this.purchasedMovies)
                data["purchasedMovies"].push(item.toJSON());
        }
        return data;
    }
}

export interface ICustomer {
    id?: string;
    name?: string;
    email?: string;
    status?: Status;
    role?: Role;
    statusExpirationDate?: Date | undefined;
    moneySpent?: number;
    purchasedMovies?: PurchasedMovie[];
}

export enum Status {
    Regular = 1,
    Advanced = 2,
}

export enum Role {
    Regular = 1,
    Admin = 2,
}

export class PurchasedMovie implements IPurchasedMovie {
    id?: string;
    movie?: Movie;
    customer?: Customer;
    purchaseDate?: Date;
    expirationDate?: Date | undefined;
    price?: number;

    constructor(data?: IPurchasedMovie) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.id = _data["id"];
            this.movie = _data["movie"] ? Movie.fromJS(_data["movie"]) : <any>undefined;
            this.customer = _data["customer"] ? Customer.fromJS(_data["customer"]) : <any>undefined;
            this.purchaseDate = _data["purchaseDate"] ? new Date(_data["purchaseDate"].toString()) : <any>undefined;
            this.expirationDate = _data["expirationDate"] ? new Date(_data["expirationDate"].toString()) : <any>undefined;
            this.price = _data["price"];
        }
    }

    static fromJS(data: any): PurchasedMovie {
        data = typeof data === 'object' ? data : {};
        let result = new PurchasedMovie();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["id"] = this.id;
        data["movie"] = this.movie ? this.movie.toJSON() : <any>undefined;
        data["customer"] = this.customer ? this.customer.toJSON() : <any>undefined;
        data["purchaseDate"] = this.purchaseDate ? this.purchaseDate.toISOString() : <any>undefined;
        data["expirationDate"] = this.expirationDate ? this.expirationDate.toISOString() : <any>undefined;
        data["price"] = this.price;
        return data;
    }
}

export interface IPurchasedMovie {
    id?: string;
    movie?: Movie;
    customer?: Customer;
    purchaseDate?: Date;
    expirationDate?: Date | undefined;
    price?: number;
}

export class Movie implements IMovie {
    id?: string;
    title?: string;
    licencingType?: LicencingType;
    description?: string;

    constructor(data?: IMovie) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.id = _data["id"];
            this.title = _data["title"];
            this.licencingType = _data["licencingType"];
            this.description = _data["description"];
        }
    }

    static fromJS(data: any): Movie {
        data = typeof data === 'object' ? data : {};
        let result = new Movie();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["id"] = this.id;
        data["title"] = this.title;
        data["licencingType"] = this.licencingType;
        data["description"] = this.description;
        return data;
    }
}
// ...
export class UpdateCustomerCommand implements IUpdateCustomerCommand {
    id?: string;
    name?: string;
    email?: string;
    role?: Role;

    constructor(data?: IUpdateCustomerCommand) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.id = _data["id"];
            this.name = _data["name"];
            this.email = _data["email"];
            this.role = _data["role"];
        }
    }

    static fromJS(data: any): UpdateCustomerCommand {
        data = typeof data === 'object' ? data : {};
        let result = new UpdateCustomerCommand();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["id"] = this.id;
        data["name"] = this.name;
        data["email"] = this.email;
        data["role"] = this.role;
        return data;
    }
}

export interface IUpdateCustomerCommand {
    id?: string;
    name?: string;
    email?: string;
    role?: Role;
}